+++ packages/client/src/pages/error/ui/ErrorLayout.module.scss
