+++ packages/client/src/pages/error/ui/index.ts
@@ -1,0 +1,2 @@
+export { NotFoundPage } from './NotFoundPage';
+export { ServerErrorPage } from './ServerErrorPage';
